[AAE-5687] clean-up form folder
