+++ src/app/stories/form/form-field/form-field.stories.ts
@@ -71,7 +71,8 @@ export default {
         { provide: SitesService, useClass: SitesServiceStub }
       ]
     })
-  ]
+  ],
+  argTypes: { limitWidth: { table: { disable: true } } }
 } as Meta;
 
 const Template: Story<FormFieldComponent> = (args) => ({
