+++ src/app/stories/form/form-renderer/form-renderer.component.html
@@ -59,53 +59,3 @@
     </mat-card>
   </div>
 </div>
-<!--
-<div *ngIf="!hasForm()">
-  <ng-content select="[empty-form]">
-  </ng-content>
-</div>
-
-<div *ngIf="hasForm()" class="adf-form-container">
-  <mat-card>
-      <mat-card-header>
-          <mat-card-title>
-              <h4>
-                  <div class="adf-form-validation-button">
-                      <i id="adf-valid-form-icon" class="material-icons"
-                          *ngIf="formDefinition.isValid; else no_valid_form">check_circle</i>
-                      <ng-template #no_valid_form>
-                          <i id="adf-invalid-form-icon" class="material-icons adf-invalid-color">error</i>
-                      </ng-template>
-                  </div>
-                  <div class="adf-form-reload-button">
-                      <button mat-icon-button (click)="onRefreshClicked()">
-                          <mat-icon>refresh</mat-icon>
-                      </button>
-                  </div>
-                  <span class="adf-form-title" [matTooltip]="formDefinition.taskName">
-                      {{formDefinition.taskName}}
-                      <ng-container *ngIf="!formDefinition.taskName">
-                          {{'FORM.FORM_RENDERER.NAMELESS_TASK' | translate}}
-                      </ng-container>
-                  </span>
-
-              </h4>
-          </mat-card-title>
-      </mat-card-header>
-      <mat-card-content>
-          <adf-form-renderer [formDefinition]="formDefinition">
-          </adf-form-renderer>
-      </mat-card-content>
-      <mat-card-actions *ngIf="formDefinition.hasOutcomes()" class="adf-form-mat-card-actions">
-          <ng-content select="adf-cloud-form-custom-outcomes"></ng-content>
-          <ng-container *ngFor="let outcome of formDefinition.outcomes">
-              <button *ngIf="outcome.isVisible" [id]="'adf-form-'+ outcome.name  | formatSpace" [color]="getColorForOutcome(outcome.name)"
-                  mat-button [disabled]="!isOutcomeButtonEnabled(outcome)"
-                  [class.adf-form-hide-button]="!isOutcomeButtonVisible(outcome, formDefinition.readOnly)"
-                  (click)="onOutcomeClicked(outcome)">
-                  {{outcome.name | translate | uppercase }}
-              </button>
-          </ng-container>
-      </mat-card-actions>
-  </mat-card>
-</div> -->
