+++ src/app/stories/form/form-renderer/form-renderer.models.ts
@@ -1117,178 +1117,6 @@ export const processServiceForm = {
   }
 };
 
-export const processServiceCloudForm = {
-  formRepresentation: {
-    id: 'text-form',
-    name: 'test-start-form',
-    version: 0,
-    description: '',
-    formDefinition: {
-      tabs: [],
-      fields: [
-        {
-          id: '1511517333638',
-          type: 'container',
-          fieldType: 'ContainerRepresentation',
-          name: 'Label',
-          tab: null,
-          numberOfColumns: 2,
-          fields: {
-            '1': [
-              {
-                fieldType: 'FormFieldRepresentation',
-                id: 'texttest',
-                name: 'texttest',
-                type: 'text',
-                value: null,
-                required: false,
-                placeholder: 'text',
-                params: {
-                  existingColspan: 2,
-                  maxColspan: 6,
-                  inputMaskReversed: true,
-                  inputMask: '0#',
-                  inputMaskPlaceholder: '(0-9)'
-                }
-              }
-            ],
-            '2': [
-              {
-                fieldType: 'AttachFileFieldRepresentation',
-                id: 'attachfiletest',
-                name: 'attachfiletest',
-                type: 'upload',
-                required: true,
-                colspan: 2,
-                placeholder: 'attachfile',
-                params: {
-                  existingColspan: 2,
-                  maxColspan: 2,
-                  fileSource: {
-                    serviceId: 'local-file',
-                    name: 'Local File'
-                  },
-                  multiple: true,
-                  link: false
-                },
-                visibilityCondition: {}
-              }
-            ]
-          }
-        }
-      ],
-      outcomes: [
-        { id: 'action1', name: 'Action 1' },
-        { id: 'action2', name: 'Action 2' }
-      ],
-      metadata: {
-        property1: 'value1',
-        property2: 'value2'
-      },
-      variables: [
-        {
-          name: 'variable1',
-          type: 'string',
-          value: 'value1'
-        },
-        {
-          name: 'variable2',
-          type: 'string',
-          value: 'value2'
-        }
-      ]
-    }
-  }
-};
-
-export const conditionalUploadWidgetsMock: any = {
-  formRepresentation: {
-    id: 'form-fb7858f7-5cf6-4afe-b462-c15a5dc0c34c',
-    name: 'AttachVisibility',
-    description: '',
-    version: 0,
-    formDefinition: {
-      tabs: [],
-      fields: [
-        {
-          id: '1dc63387-aa9d-4f06-adfa-37817e8fd394',
-          name: 'Label',
-          type: 'container',
-          tab: null,
-          numberOfColumns: 2,
-          fields: {
-            '1': [
-              {
-                id: 'Text0xlk8n',
-                name: 'Text',
-                type: 'text',
-                required: false,
-                colspan: 1,
-                placeholder: null,
-                minLength: 0,
-                maxLength: 0,
-                regexPattern: null,
-                visibilityCondition: null,
-                params: {
-                  existingColspan: 1,
-                  maxColspan: 2
-                }
-              }
-            ],
-            '2': [
-              {
-                id: 'Attachfile0h9fr1',
-                name: 'Attach file',
-                type: 'upload',
-                required: false,
-                colspan: 1,
-                visibilityCondition: {
-                  leftFormFieldId: 'Text0xlk8n',
-                  leftRestResponseId: '',
-                  operator: '==',
-                  rightValue: 'Attach',
-                  rightType: null,
-                  rightFormFieldId: '',
-                  rightRestResponseId: '',
-                  nextConditionOperator: '',
-                  nextCondition: null
-                },
-                params: {
-                  existingColspan: 1,
-                  maxColspan: 2,
-                  fileSource: {
-                    serviceId: 'alfresco-content',
-                    name: 'Alfresco Content'
-                  },
-                  multiple: false,
-                  link: false
-                }
-              }
-            ]
-          }
-        }
-      ],
-      outcomes: [
-        {
-          id: '5f2f1c2d-5a79-4ed1-a262-4fef190d41eb',
-          name: 'Custom Outcome',
-          visibilityCondition: {
-            leftType: 'field',
-            leftValue: 'Text0xlk8n',
-            operator: '==',
-            rightValue: 'hi',
-            rightType: 'value',
-            nextConditionOperator: '',
-            nextCondition: null
-          }
-        }
-      ],
-      metadata: {},
-      variables: []
-    }
-  }
-};
-
 export const cloudFormMock = {
   id: 'form-b661635a-dc3e-4557-914a-3498ed47189c',
   name: 'form-with-all-fields',
@@ -2074,368 +1902,4 @@ export const cloudFormMock = {
       value: ''
     }
   ]
-};
-
-export let fakeMetadataForm = {
-  id: 'form-de8895be-d0d7-4434-beef-559b15305d72',
-  name: 'StartEventForm',
-  description: '',
-  version: 0,
-  formDefinition: {
-    tabs: [],
-    fields: [
-      {
-        type: 'container',
-        id: '5a6b24c1-db2b-45e9-9aff-142395433d23',
-        name: 'Label',
-        tab: null,
-        fields: {
-          '1': [
-            {
-              type: 'text',
-              id: 'pfx_property_one',
-              name: 'pfx_property_one',
-              colspan: 1,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2
-              },
-              visibilityCondition: null,
-              placeholder: null,
-              value: null,
-              required: false,
-              minLength: 0,
-              maxLength: 0,
-              regexPattern: null
-            }
-          ],
-          '2': [
-            {
-              type: 'boolean',
-              id: 'pfx_property_two',
-              name: 'pfx_property_two',
-              colspan: 1,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2
-              },
-              visibilityCondition: null,
-              placeholder: null,
-              value: null,
-              required: false,
-              minLength: 0,
-              maxLength: 0,
-              regexPattern: null
-            }
-          ],
-          '3': [
-            {
-              id: 'content_form_nodes',
-              name: 'Nodes',
-              type: 'upload',
-              readOnly: false,
-              required: true,
-              colspan: 1,
-              visibilityCondition: null,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2,
-                fileSource: {
-                  serviceId: 'alfresco-content',
-                  name: 'Alfresco Content',
-                  metadataAllowed: true
-                },
-                multiple: true,
-                menuOptions: {
-                  show: true,
-                  download: true,
-                  retrieveMetadata: true,
-                  remove: true
-                },
-                link: false
-              }
-            }
-          ],
-          '4': [
-            {
-              id: 'pfx_property_three',
-              name: 'pfx_property_three',
-              required: false,
-              readOnly: false,
-              colspan: 1,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2
-              },
-              visibilityCondition: null,
-              type: 'dropdown',
-              optionType: 'manual',
-              options: [
-                {
-                  id: 'empty',
-                  name: 'Choose one...'
-                },
-                {
-                  id: 'opt_1',
-                  name: 'Option 1'
-                },
-                {
-                  id: 'opt_2',
-                  name: 'Option 2'
-                }
-              ],
-              value: 'empty',
-              restUrl: null,
-              restResponsePath: null,
-              restIdProperty: null,
-              restLabelProperty: null
-            }
-          ],
-          '5': [
-            {
-              id: 'pfx_property_four',
-              name: 'pfx_property_four',
-              required: false,
-              readOnly: false,
-              colspan: 1,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2
-              },
-              visibilityCondition: null,
-              type: 'dropdown',
-              optionType: 'manual',
-              options: [
-                {
-                  id: 'empty',
-                  name: 'Choose one...'
-                },
-                {
-                  id: 'option_1',
-                  name: 'Option: 1'
-                },
-                {
-                  id: 'option_2',
-                  name: 'Option: 2'
-                }
-              ],
-              value: 'empty',
-              restUrl: null,
-              restResponsePath: null,
-              restIdProperty: null,
-              restLabelProperty: null
-            }
-          ],
-          '6': [
-            {
-              id: 'pfx_property_five',
-              name: 'pfx_property_five',
-              required: false,
-              readOnly: false,
-              colspan: 1,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2
-              },
-              visibilityCondition: null,
-              type: 'dropdown',
-              optionType: 'manual',
-              options: [
-                {
-                  id: 'empty',
-                  name: 'Choose one...'
-                },
-                {
-                  id: 'green',
-                  name: 'Colour green'
-                },
-                {
-                  id: 'orange',
-                  name: 'Colour orange'
-                }
-              ],
-              value: 'empty',
-              restUrl: null,
-              restResponsePath: null,
-              restIdProperty: null,
-              restLabelProperty: null
-            }
-          ],
-          '7': [
-            {
-              id: 'cmfb85b2a7295ba41209750bca176ccaf9a',
-              name: 'File viewer',
-              type: 'file-viewer',
-              readOnly: false,
-              required: false,
-              colspan: 1,
-              visibilityCondition: null,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2,
-                uploadWidget: 'content_form_nodes'
-              }
-            }
-          ],
-          '8': [
-            {
-              type: 'text',
-              id: 'pfx_property_six',
-              name: 'pfx_property_six',
-              colspan: 1,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2
-              },
-              visibilityCondition: null,
-              placeholder: null,
-              value: null,
-              required: false,
-              minLength: 0,
-              maxLength: 0,
-              regexPattern: null
-            }
-          ],
-          '9': [
-            {
-              type: 'text',
-              id: 'pfx_property_seven',
-              name: 'pfx_property_seven',
-              colspan: 1,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2
-              },
-              visibilityCondition: null,
-              placeholder: null,
-              value: null,
-              required: false,
-              minLength: 0,
-              maxLength: 0,
-              regexPattern: null
-            }
-          ],
-          '10': [
-            {
-              type: 'text',
-              id: 'pfx_property_eight',
-              name: 'pfx_property_eight',
-              colspan: 1,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2
-              },
-              visibilityCondition: null,
-              placeholder: null,
-              value: null,
-              required: false,
-              minLength: 0,
-              maxLength: 0,
-              regexPattern: null
-            }
-          ]
-        },
-        numberOfColumns: 2
-      }
-    ],
-    outcomes: [],
-    metadata: {},
-    variables: []
-  }
-};
-
-export let fakeViewerForm = {
-  id: 'form-de8895be-d0d7-4434-beef-559b15305d72',
-  name: 'StartEventForm',
-  description: '',
-  version: 0,
-  formDefinition: {
-    tabs: [],
-    fields: [
-      {
-        type: 'container',
-        id: '5a6b24c1-db2b-45e9-9aff-142395433d23',
-        name: 'Label',
-        tab: null,
-        fields: {
-          '1': [
-            {
-              id: 'content_form_nodes',
-              name: 'Nodes',
-              type: 'upload',
-              readOnly: false,
-              required: true,
-              colspan: 1,
-              visibilityCondition: null,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2,
-                fileSource: {
-                  serviceId: 'alfresco-content',
-                  name: 'Alfresco Content',
-                  metadataAllowed: true
-                },
-                multiple: true,
-                menuOptions: {
-                  show: true,
-                  download: true,
-                  retrieveMetadata: true,
-                  remove: true
-                },
-                link: false
-              }
-            }
-          ],
-          '2': [
-            {
-              id: 'upload_widget',
-              name: 'Nodes',
-              type: 'upload',
-              readOnly: false,
-              required: true,
-              colspan: 1,
-              visibilityCondition: null,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2,
-                fileSource: {
-                  serviceId: 'alfresco-content',
-                  name: 'Alfresco Content',
-                  metadataAllowed: true
-                },
-                multiple: true,
-                menuOptions: {
-                  show: true,
-                  download: true,
-                  retrieveMetadata: true,
-                  remove: true
-                },
-                link: false
-              }
-            }
-          ],
-          '3': [
-            {
-              id: 'cmfb85b2a7295ba41209750bca176ccaf9a',
-              name: 'File viewer',
-              type: 'file-viewer',
-              readOnly: false,
-              required: false,
-              colspan: 1,
-              visibilityCondition: null,
-              params: {
-                existingColspan: 1,
-                maxColspan: 2,
-                uploadWidget: 'content_form_nodes'
-              }
-            }
-          ]
-        },
-        numberOfColumns: 2
-      }
-    ],
-    outcomes: [],
-    metadata: {},
-    variables: []
-  }
 };
