+++ src/app/stories/form/form-renderer/visibility-condition-task/task-form/task-detail.models.ts
@@ -1,6 +1,6 @@
 import { TaskDetailsModel } from '@alfresco/adf-process-services';
 
-export let standaloneTaskWithForm = new TaskDetailsModel({
+export const standaloneTaskWithForm = new TaskDetailsModel({
   id: '100',
   name: 'Standalone Task With Form',
   description: null,
@@ -28,46 +28,6 @@ export let standaloneTaskWithForm = new TaskDetailsModel({
   processDefinitionVersion: null,
   processDefinitionDeploymentId: null,
   formKey: '222',
-  processInstanceStartUserId: null,
-  initiatorCanCompleteTask: false,
-  adhocTaskCanBeReassigned: false,
-  taskDefinitionKey: 'sid-DDECD9E4-0299-433F-9193-C3D905C3EEBE',
-  executionId: '86',
-  involvedGroups: [],
-  involvedPeople: [],
-  memberOfCandidateUsers: false,
-  managerOfCandidateGroup: false,
-  memberOfCandidateGroup: false
-});
-
-export let standaloneTaskWithoutForm = new TaskDetailsModel({
-  id: '200',
-  name: 'Standalone Task Without Form',
-  description: null,
-  category: null,
-  assignee: {
-    id: 1001,
-    firstName: 'Wilbur',
-    lastName: 'Adams',
-    email: 'wilbur@app.activiti.com'
-  },
-  created: '2016-11-03T15:25:42.749+0000',
-  dueDate: null,
-  endDate: null,
-  duration: null,
-  priority: 50,
-  parentTaskId: null,
-  parentTaskName: null,
-  processInstanceId: null,
-  processInstanceName: null,
-  processDefinitionId: null,
-  processDefinitionName: null,
-  processDefinitionDescription: null,
-  processDefinitionKey: null,
-  processDefinitionCategory: null,
-  processDefinitionVersion: null,
-  processDefinitionDeploymentId: null,
-  formKey: null,
   processInstanceStartUserId: null,
   initiatorCanCompleteTask: false,
   adhocTaskCanBeReassigned: false,
