+++ src/app/stories/form/mock/stub-services.ts
@@ -1,6 +1,5 @@
 import {
   AppConfigService,
-  BpmUserModel,
   ContentLinkModel,
   FormEvent,
   GroupModel,
@@ -15,8 +14,8 @@ import { MinimalNode, NodeEntry, SitePaging, UserRepresentation } from '@alfresc
 import { Injectable } from '@angular/core';
 import { from, Observable, of, ReplaySubject, Subject } from 'rxjs';
 import { standaloneTaskWithForm } from '../form-renderer/visibility-condition-task/task-form/task-detail.models';
+import { fakeBpmUser, mockIdentityGroups, mockIdentityUsers } from './fake-users-groups';
 import {
-  AlfrescoApiActivitiStub,
   AuthenticationApiStub,
   ContentApiStub,
   GroupApiStub,
@@ -27,28 +26,6 @@ import {
   TaskApiStub,
   TaskFormsApiStub
 } from './stub-apis';
-
-const fakeBpmUser = new BpmUserModel({
-  apps: [],
-  capabilities: null,
-  company: 'fake-company',
-  created: 'fake-create-date',
-  email: 'fakeBpm@fake.com',
-  externalId: 'fake-external-id',
-  firstName: 'fake-bpm-first-name',
-  lastName: 'fake-bpm-last-name',
-  groups: [],
-  id: 'fake-id',
-  lastUpdate: 'fake-update-date',
-  latestSyncTimeStamp: 'fake-timestamp',
-  password: 'fake-password',
-  pictureId: 12,
-  status: 'fake-status',
-  tenantId: 'fake-tenant-id',
-  tenantName: 'fake-tenant-name',
-  tenantPictureId: 'fake-tenant-picture-id',
-  type: 'fake-type'
-});
 
 export class AuthenticationServiceStub {
   onLogin: ReplaySubject<any> = new ReplaySubject<any>(1);
@@ -93,47 +70,6 @@ export class AlfrescoApiServiceStub {
   }
 
   getInstance = () => this;
-
-  constructor(private appConfig: AppConfigService) {}
-}
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AlfrescoApiCompatibilityServiceStub {
-  ///test
-
-  nodesApi = new NodesApiStub();
-
-  nodes = this.nodesApi;
-
-  taskApi = new TaskApiStub();
-
-  activiti = { taskApi: new TaskApiStub(), taskFormsApi: new TaskFormsApiStub(), contentApi: new ContentApiStub(), processApi: new ProcessApiStub() };
-
-  nodeUpdated = new Subject<Node>();
-
-  modelsApi = new ModelsApiStub();
-
-  contentApi = new ContentApiStub();
-
-  content = new ContentApiStub();
-
-  authenticationApi = new AuthenticationApiStub();
-
-  alfrescoApiInitialized: ReplaySubject<boolean> = new ReplaySubject(1);
-
-  alfrescoApi = {
-    activiti: {
-      alfrescoApi: new AlfrescoApiActivitiStub()
-    }
-  };
-
-  async load() {
-    await this.appConfig.load();
-  }
-
-  getInstance = () => this.alfrescoApi;
 
   constructor(private appConfig: AppConfigService) {}
 }
@@ -232,44 +168,6 @@ export class CustomResourcesServiceStub {
   }
 }
 
-export const mockIdentityUser1: IdentityUserModel = {
-  id: 'mock-user-id-1',
-  username: 'userName1',
-  firstName: 'first-name-1',
-  lastName: 'last-name-1',
-  email: 'abc@xyz.com'
-};
-export const mockIdentityUser2: IdentityUserModel = {
-  id: 'mock-user-id-2',
-  username: 'userName2',
-  firstName: 'first-name-2',
-  lastName: 'last-name-2',
-  email: 'abcd@xyz.com'
-};
-export const mockIdentityUser3: IdentityUserModel = {
-  id: 'mock-user-id-3',
-  username: 'userName3',
-  firstName: 'first-name-3',
-  lastName: 'last-name-3',
-  email: 'abcde@xyz.com'
-};
-export const mockIdentityUser4: IdentityUserModel = {
-  id: 'mock-user-id-4',
-  username: 'userName4',
-  firstName: 'first-name-4',
-  lastName: 'last-name-4',
-  email: 'abcde@xyz.com'
-};
-export let mockIdentityUser5: IdentityUserModel = {
-  id: 'mock-user-id-5',
-  username: 'userName5',
-  firstName: 'first-name-5',
-  lastName: 'last-name-5',
-  email: 'abcde@xyz.com'
-};
-
-export const mockIdentityUsers: IdentityUserModel[] = [mockIdentityUser1, mockIdentityUser2, mockIdentityUser3, mockIdentityUser4, mockIdentityUser5];
-
 export class IdentityUserServiceStub {
   findUsersByName(search: string): Observable<IdentityUserModel[]> {
     if (search === '') {
@@ -282,45 +180,7 @@ export class IdentityUserServiceStub {
   getClientIdByApplicationName(_applicationName: string): Observable<string> {
     return from('fakeUser');
   }
-  //getCurrentUserInfo = (): IdentityUserModel => this.fakeIdentityUser;
 }
-
-export const mockIdentityGroup1 = <IdentityGroupModel>{
-  id: 'mock-group-id-1',
-  name: 'Mock Group 1',
-  path: '/mock',
-  subGroups: []
-};
-
-export const mockIdentityGroup2 = <IdentityGroupModel>{
-  id: 'mock-group-id-2',
-  name: 'Mock Group 2',
-  path: '',
-  subGroups: []
-};
-
-export const mockIdentityGroup3 = <IdentityGroupModel>{
-  id: 'mock-group-id-3',
-  name: 'Mock Group 3',
-  path: '',
-  subGroups: []
-};
-
-export const mockIdentityGroup4 = <IdentityGroupModel>{
-  id: 'mock-group-id-4',
-  name: 'Mock Group 4',
-  path: '',
-  subGroups: []
-};
-
-export const mockIdentityGroup5 = <IdentityGroupModel>{
-  id: 'mock-group-id-5',
-  name: 'Mock Group 5',
-  path: '',
-  subGroups: []
-};
-
-export const mockIdentityGroups = [mockIdentityGroup1, mockIdentityGroup2, mockIdentityGroup3, mockIdentityGroup4, mockIdentityGroup5];
 
 export class IdentityGroupServiceStub {
   findGroupsByName(searchParams: IdentityGroupSearchParam): Observable<IdentityGroupModel[]> {
@@ -334,44 +194,4 @@ export class IdentityGroupServiceStub {
   getClientIdByApplicationName(_applicationName: string): Observable<string> {
     return from('fakeUser');
   }
-  //getCurrentUserInfo = (): IdentityUserModel => this.fakeIdentityUser;
 }
-
-// export class ProcessCloudContentServiceStub {
-
-//   constructor(private contentService: ContentService, private downloadService: DownloadService){
-//   }
-
-//   async downloadFile(nodeId: string) {
-
-//     const ticket = await this.getAuthTicket();
-//     const url = this.contentService.getContentUrl(nodeId, true, ticket);
-
-//     this.downloadService.downloadUrl(url, nodeId);
-//   }
-
-//   getAuthTicket(): Promise<string> {
-//     return Promise.resolve('ticket')
-//   }
-// }
-
-// export class ProcessContentServiceStub {
-
-//   contentApi = new ContentApiStub()
-
-//   getFileContent(_contentId: number): Observable<RelatedContentRepresentation> {
-//     return of(fakePngAnswer)
-//   }
-
-//   getFileRawContentUrl(contentId: number): string {
-//     return this.contentApi.getRawContentUrl(contentId)
-// }
-
-// getFileRawContent(contentId: number): Observable<Blob>{
-//     return from(this.contentApi.getRawContent(contentId));
-//   }
-
-//   getContentPreview(contentId: number): Observable<Blob> {
-//     return from(this.contentApi.getContentPreview(contentId))
-//   }
-// }
